rename FormatTags --> FormatMessage
