+++ web/src/lib/components/admin-page/settings/machine-learning-settings/machine-learning-settings.svelte
@@ -13,7 +13,7 @@
   import SettingSwitch from '$lib/components/shared-components/settings/setting-switch.svelte';
   import { featureFlags } from '$lib/stores/server-config.store';
   import { json, t } from 'svelte-i18n';
-  import FormatTags from '$lib/components/i18n/format-tags.svelte';
+  import FormatMessage from '$lib/components/i18n/format-message.svelte';
 
   export let savedConfig: SystemConfigDto;
   export let defaultConfig: SystemConfigDto;
@@ -71,9 +71,9 @@
             isEdited={config.machineLearning.clip.modelName !== savedConfig.machineLearning.clip.modelName}
           >
             <p slot="desc" class="immich-form-label pb-2 text-sm">
-              <FormatTags message={$json('admin.machine_learning_clip_model_description')} let:message>
+              <FormatMessage message={$json('admin.machine_learning_clip_model_description')} let:message>
                 <a href="https://huggingface.co/immich-app"><u>{message}</u></a>
-              </FormatTags>
+              </FormatMessage>
             </p>
           </SettingInputField>
         </div>
