+++ web/src/lib/components/i18n/format-message.svelte
